Setup versioned routes
and pull shared stuff out of routes
